changed process_axis to process_axes and improved docs
