+++ easy_mpl/__init__.py
@@ -5,9 +5,9 @@ from ._bar import bar_chart
 from ._regplot import regplot
 from ._imshow import imshow
 from ._hist import hist
-from ._main import scatter
+from ._scatter import scatter
 from ._contour import contour
-from ._main import dumbbell_plot
+from ._dumbell import dumbbell_plot
 from ._ridge import ridge
 from ._parallel_coords import parallel_coordinates
 from ._taylor_plot import taylor_plot
@@ -18,7 +18,7 @@ from ._box import boxplot
 from ._violin import violin_plot
 
 
-from .utils import process_axis
+from .utils import process_axes
 
 
 __version__ = "0.21.1"
