+++ easy_mpl/_main.py
@@ -1,23 +1,16 @@
 
-__all__ = [
-    "plot",
-    "scatter",
-    "dumbbell_plot"
-]
-
-from typing import Tuple
+__all__ = ["plot"]
 
 import numpy as np
-import matplotlib as mpl
 import matplotlib.pyplot as plt
-import matplotlib.lines as mlines
 
-from .utils import to_1d_array, process_axis
+from .utils import process_axes
 
 
 def plot(
         *args,
         show: bool = True,
+        ax: plt.Axes = None,
         **kwargs
 ) -> plt.Axes:
     """
@@ -31,14 +24,17 @@ def plot(
         *args :
             either a single array or x and y arrays or anything which can go to
             `axes.plot()`_ or anything which can got to `plt.plot()`_ .
+        ax : :obj:`matplotlib.axes`
+            matplotlib axes object on which plot is to be drawn. If not given,
+            then current active axes will be used.
         show : bool, optional
         **kwargs : optional
             Anything which goes into `easy_mpl.utils.process_axis`.
 
     Returns
     -------
-    matplotlib.pyplot.Axes
-        matplotlib Axes on which the plot is drawn. If ``show`` is False, this axes
+    :obj:`matplotlib.axes`
+        :obj:`matplotlib.axes` on which the plot is drawn. If ``show`` is False, this axes
         can be used for further processing
 
     Example
@@ -103,13 +99,7 @@ def plot(
         plot_args.append(marker)
         assert 'marker' not in _plot_kwargs  # we have alreay got marker
 
-    if 'ax' in kwargs:
-        ax = kwargs.pop('ax')
-    else:
-        ax = plt.gca()
-
-    if ax is None:
-        # it is possible that as is given in kwargs but given as None
+    if not ax:
         ax = plt.gca()
 
     if 'figsize' in kwargs:
@@ -145,201 +135,10 @@ def plot(
 
 def _process_axis(ax, show, kwargs):
     if kwargs:
-        ax = process_axis(ax=ax, **kwargs)
+        ax = process_axes(ax=ax, **kwargs)
     if kwargs.get('save', False):
         plt.savefig(f"{kwargs.get('name', 'fig.png')}")
     if show:
         plt.show()
     return ax
 
-
-def scatter(
-        x,
-        y,
-        colorbar: bool = False,
-        colorbar_orientation: str = "vertical",
-        show: bool = True,
-        ax: plt.Axes = None,
-        **kwargs
-) -> Tuple[plt.Axes, mpl.collections.PathCollection]:
-    """
-    scatter plot between two arrays x and y
-
-    Parameters
-    ----------
-    x : list, array
-    y : list, array
-    colorbar : bool, optional
-    colorbar_orientation : str, optional
-    show : bool, optional
-        whether to show the plot or not
-    ax : plt.Axes, optional
-        if not given, current available axes will be used
-    **kwargs : optional
-        any additional keyword arguments for `axes.scatter`_
-
-    Returns
-    --------
-    tuple :
-        A tuple whose first member is matplotlib Axes and second member is
-    matplotlib.collections.PathCollection
-
-    Examples
-    --------
-        >>> from easy_mpl import scatter
-        >>> import numpy as np
-        >>> import matplotlib.pyplot as plt
-        >>> x_ = np.random.random(100)
-        >>> y_ = np.random.random(100)
-        >>> scatter(x_, y_, show=False)
-        ... # show colorbar
-        >>> scatter(x_, y_, colorbar=True, show=False)
-        ... # retrieve axes for further processing
-        >>> axes, _ = scatter(x_, y_, show=False)
-        >>> assert isinstance(axes, plt.Axes)
-
-    See :ref:`sphx_glr_auto_examples_scatter.py` for more examples
-
-    .. _axes.scatter:
-        https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.scatter.html
-
-    """
-    if ax is None:
-        ax = plt.gca()
-        if 'figsize' in kwargs:
-            figsize = kwargs.pop('figsize')
-            ax.figure.set_size_inches(figsize)
-
-    x = to_1d_array(x)
-    y = to_1d_array(y)
-
-    if colorbar and 'c' not in kwargs:
-        kwargs['c'] = np.arange(len(x))
-
-    sc = ax.scatter(x, y, **kwargs)
-
-    if colorbar:
-        fig: plt.Figure = plt.gcf()
-        fig.colorbar(sc, orientation=colorbar_orientation, pad=0.1)
-
-    if show:
-        plt.show()
-
-    return ax, sc
-
-
-def dumbbell_plot(
-        start,
-        end,
-        labels=None,
-        start_kws: dict = None,
-        end_kws: dict = None,
-        line_kws: dict = None,
-        show: bool = True,
-        ax: plt.Axes = None,
-        **kwargs
-) -> plt.Axes:
-    """
-    Dumbell plot which indicates variation of several variables
-    from start to end.
-
-    Parameters
-    ----------
-        start : list, array, series
-            an array consisting of starting values
-        end : list, array, series
-            an array consisting of end values
-        labels : list, array, series, optional
-            names of values in start/end arrays. It is used to label
-            ticklabcls on y-axis
-        start_kws : dict, optional
-            any additional keyword arguments for `axes.scatter`_ to modify start
-            markers such as ``color``, ``label`` etc
-        end_kws : dict, optional
-            any additional keyword arguments for `axes.scatter`_ to modify end
-            markers such as ``color``, ``label`` etc
-        line_kws : dict, optional
-            any additional keyword arguments for `lines.Line2D`_ to modify line
-            style/color which connects dumbbells.
-        show : bool, optional
-            whether to show the plot or not
-        ax : plt.Axes, optional
-            matplotlib axes object to work with. If not given then currently available
-            axes will be used.
-        **kwargs :
-            any additional keyword arguments for `process_axis`.
-
-    Returns
-    -------
-        matplotlib Axes object.
-
-    Examples
-    --------
-        >>> import numpy as np
-        >>> from easy_mpl import dumbbell_plot
-        >>> st = np.random.randint(1, 5, 10)
-        >>> en = np.random.randint(11, 20, 10)
-        >>> dumbbell_plot(st, en)
-        ... # modify line color
-        >>> dumbbell_plot(st, en, line_kws={'color':"black"})
-
-    See :ref:`sphx_glr_auto_examples_dumbell.py` for more examples
-
-    .. _axes.scatter:
-        https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.scatter.html
-
-    .. _lines.Line2D:
-        https://matplotlib.org/stable/api/_as_gen/matplotlib.lines.Line2D.html
-
-    """
-    if ax is None:
-        ax = plt.gca()
-        if 'figsize' in kwargs:
-            figsize = kwargs.pop('figsize')
-            ax.figure.set_size_inches(figsize)
-
-    # convert starting and ending values to 1d array
-    start = to_1d_array(start)
-    end = to_1d_array(end)
-
-    index = np.arange(len(start))
-
-    assert len(start) == len(end) == len(index)
-
-    if labels is None:
-        labels = np.arange(len(index))
-
-    line_kws = line_kws or {'color': 'skyblue'}
-
-    # draw line segment
-    def lien_segment(p1, p2, axes):
-        l = mlines.Line2D([p1[0], p2[0]], [p1[1], p2[1]], **line_kws)
-        axes.add_line(l)
-        return
-
-    # assigning colors
-    start_kws = start_kws or {'color': '#a3c4dc', "label": "start"}
-    end_kws = end_kws or {'color': '#0e668b', "label": "end"}
-
-    # plotting points for starting and ending values
-    ax, _ = scatter(y=index, x=start, show=False, ax=ax, **start_kws)
-    ax, _ = scatter(y=index, x=end, ax=ax, show=False, **end_kws)
-
-    ax.legend()
-
-    # joining points together using line segments
-    for idx, _p1, _p2 in zip(index, end, start):
-        lien_segment([_p1, idx], [_p2, idx], ax)
-
-    # set labels
-    ax.set_yticks(index)
-    ax.set_yticklabels(labels)
-
-    if kwargs:
-        process_axis(ax=ax, **kwargs)
-    # show plot if show=True
-    if show:
-        plt.tight_layout()  # todo should we put it outside of if?
-        plt.show()
-
-    return ax
